move helpful raises_exception to general module
